backend/requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
email-validator==2.2.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
flask-openapi3[swagger]==4.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pg8000==1.31.5
pydantic==2.12.5
PyJWT==2.10.1
python-dotenv==1.2.1
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.4

pytest==9.0.1
pytest-flask==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
Flask-Admin==2.0.0
Flask-Babel==4.0.0
